moved search bar files around
