+++ react-app/src/components/Dashboard/index.js
@@ -5,6 +5,7 @@ import { loadUserShoppingLists } from "../../store/shoppingList"
 import ShoppingListForm from './forms/ShoppingListForm'
 import ShoppingListDisplay from './ShoppingListDisplay'
 import './Dashboard.css'
+import { Link } from 'react-router-dom'
 
 export default function Dashboard() {
   const dispatch = useDispatch()
@@ -30,6 +31,7 @@ export default function Dashboard() {
       <h1>hi {user && user.username}!</h1>
       {showShoppingListForm ? <ShoppingListForm setShowForm={setShowShoppingListForm}/> : <button className="stdbutton" onClick={revealForm}>New Shopping List</button>}
       <ShoppingListDisplay />
+      <Link to={`/user/${user.id}/pantry`} >Your  Pantry</Link>
     </>
   )
 }
