+++ react-app/src/components/SearchBar/SearchBar.css
