+++ react-app/src/components/SearchBar/SearchBar.js
@@ -2,9 +2,9 @@ import { useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { searchItems } from  '../../store/items'
 import SearchResult from './SearchResult'
-import './styles/SearchBar.css'
+import './SearchBar.css'
 
-export default function SearchBar() {
+export default function SearchBar({pantry}) {
   const dispatch = useDispatch()
   const [query, setQuery] = useState("")
   const results = useSelector(state => state.items.results)
@@ -24,7 +24,7 @@ export default function SearchBar() {
     <div className="SearchBar">
       <input onChange={handleQuery}></input>
       <ul className="searchResults">
-        {results && results.map(result => <SearchResult key ={result.id} item={result} />)}
+        {results && results.map(result => <SearchResult key ={result.id} item={result} pantry={pantry}/>)}
       </ul>
     </div>
   )
