+++ react-app/src/components/SearchBar/SearchResult.js
@@ -1,15 +1,17 @@
 import { useState } from 'react';
 import { Modal } from '../../context/Modal';
-import ShoppingListItem from './forms/ShoppingListItem';
+import ShoppingListItem from '../ShoppingList/forms/ShoppingListItem';
 
-export default function SearchResult ({item}) {
+export default function SearchResult ({item, pantry}) {
   const [showModal, setShowModal] = useState(false)
   return (
     <>
       <li onClick={() => setShowModal(true)}>{item.name}</li>
       {showModal &&
         <Modal onClose={() => setShowModal(false)}>
-          <ShoppingListItem item={item} setShowModal={setShowModal}/>
+          {pantry ?
+          <ShoppingListItem item={item} setShowModal={setShowModal}/> : null
+          }
         </Modal>
       }
     </>
